First commit after a long break

+++ lib/Oxiana/Controller/Image.pm
@@ -28,29 +28,34 @@ sub index :Path :Args(0) {
     my $img = $c->req->params->{img}; $img =~ s/^$re//;
 
     if ($img) {
-	$c->res->status(500);
-	$c->detach('Modal', 'index', [ 'error' ]);
+	my $m = $c->model('Maps::BookItem');
+	my $p = $c->req->params;
+	my $item = $m->create({ item_type => 'map', poi_id => $p->{poi_id}, content => $img, book_id => $p->{book_id}, chapter_id => $p->{chapter_id} }); 
+	unless ($item) {	
+	    $c->res->status(500);
+	    $c->detach('Modal', 'index', [ 'error' ]);
+	}
+	$c->res->body('yeah');
     } else {
     
 	# my $cmd = printf "echo \"%s\" | base64 -D > foobar.png", $img;
 	
 	# $c->log->info(qx/$cmd/);
-	my $m = $c->model('Maps::BookItem');
-	my $pois = $m->search(poi_id => $c->req->params->{poi_id});
-	my $poi;
-	for ($pois->count) {
-	    /^0$/ && do {
-		# ask for chapter
-		last;
-	    };
-	    /^1$/ && do {
-		$poi = $pois->first;
-		last;
-	    };
-	    # ask for book and chapter
-	}
+
+	# my $pois = $m->search(poi_id => $c->req->params->{poi_id});
+	# my $poi;
+	# for ($pois->count) {
+	#     /^0$/ && do {
+	# 	# ask for chapter
+	# 	last;
+	#     };
+	#     /^1$/ && do {
+	# 	$poi = $pois->first;
+	# 	last;
+	#     };
+	#     # ask for book and chapter
+	# }
 	
-	# my $item = $m->create({ item_type => 'map', poi_id => $poi->id, content => $img, book_id => $poi->book_id, chapter_id => $poi->chapter_id }); 
 	
 	# check that the book exists
 	# check whether the poi is in a chapter, and if so, add
